bug fix coll comm
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@2820 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/parallel/DomainDecomposition.h
@@ -181,7 +181,7 @@ public:
 	}
 
 	float collCommGetFloat() {
-		return _collComm.getInt();
+		return _collComm.getFloat();
 	}
 
 	double collCommGetDouble() {
